pyproject.toml
[tool.poetry]
name = "android"
version = "0.1.0"
description = ""
authors = ["Magniquick <67406856+Magniquick@users.noreply.github.com>"]
license = "GPL v3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
cryptography = "^43.0.0"
loguru = "^0.7.2"
lxml = "^5.3.0"
types-requests = "^2.32.0.20241016"
lxml-stubs = "^0.5.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
